backend/requirements.txt
boto3==1.10.26
botocore==1.13.26
certifi==2019.9.11
chardet==3.0.4
Click==7.0
DateTime==4.3
dnspython==1.16.0
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==20.0.2
idna==2.8
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
polling==0.3.1
pymongo==3.9.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
serial==0.0.97
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
zope.interface==4.7.1
fsspec==0.5.2
future==0.17.1
gevent==1.4.0
glob2==0.7
google-api-core==1.16.0
google-auth==1.11.2
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-language==1.3.0
google-api-python-client==1.7.11
google-cloud-speech==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpcio==1.25.0